update code, fix install bug
