+++ BlinkerAdapters/__init__.py
@@ -1,0 +1,2 @@
+from .BlinkerLinuxWS import *
+from .BlinkerMQTT import *
